[release/2023-07-11a] Fix-forward: changelog, readme and reporting
