+++ changelog/2023-07-11a.md
@@ -1,0 +1,2 @@
+- NRLF-605 - Update Readme
+- NRLF-506 - [MI/BI] DB schema and querying the DB
